Remove EvoModelParams from DNASubstModel
