+++ phylo/src/substitution_models/dna_models/mod.rs
@@ -1,9 +1,10 @@
 use log::{info, warn};
 
 use crate::alphabets::NUCLEOTIDE_INDEX;
-use crate::evolutionary_models::{DNAModelType, EvoModelParams};
+use crate::evolutionary_models::DNAModelType;
 use crate::substitution_models::{
-    FreqVector, SubstLikelihoodCost, SubstMatrix, SubstModel, SubstModelInfo, SubstitutionModel,
+    DNAParameter::*, FreqVector, SubstLikelihoodCost, SubstMatrix, SubstModel, SubstModelInfo,
+    SubstitutionModel,
 };
 use crate::Result;
 
@@ -64,15 +65,38 @@ impl SubstitutionModel for DNASubstModel {
         &self.q
     }
 
-    fn parameter_definition(&self) -> Vec<(&'static str, Vec<Self::Parameter>)> {
-        self.params.parameter_definition()
+    fn parameter_definition(&self) -> Vec<(&'static str, Vec<DNAParameter>)> {
+        match self.params.model_type {
+            DNAModelType::JC69 => vec![],
+            DNAModelType::K80 => vec![
+                ("alpha", vec![Rtc, Rag]),
+                ("beta", vec![Rta, Rtg, Rca, Rcg]),
+            ],
+            DNAModelType::HKY => vec![
+                ("alpha", vec![Rtc, Rag]),
+                ("beta", vec![Rta, Rtg, Rca, Rcg]),
+            ],
+            DNAModelType::TN93 => vec![
+                ("alpha1", vec![Rtc]),
+                ("alpha2", vec![Rag]),
+                ("beta", vec![Rta, Rtg, Rca, Rcg]),
+            ],
+            DNAModelType::GTR => vec![
+                ("rca", vec![Rca]),
+                ("rcg", vec![Rcg]),
+                ("rta", vec![Rta]),
+                ("rtc", vec![Rtc]),
+                ("rtg", vec![Rtg]),
+            ],
+            _ => unreachable!(),
+        }
     }
 
-    fn param(&self, param_name: &Self::Parameter) -> f64 {
+    fn param(&self, param_name: &DNAParameter) -> f64 {
         self.params.param(param_name)
     }
 
-    fn set_param(&mut self, param_name: &Self::Parameter, value: f64) {
+    fn set_param(&mut self, param_name: &DNAParameter, value: f64) {
         self.params.set_param(param_name, value);
         self.update();
     }
